+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/gui/widget/PackStructureWidget.kt
@@ -1,0 +1,4 @@
+package liltojustice.trueadaptivemusic.client.gui.widget
+
+class PackStructureWidget {
+}
